requirements.txt
flask==3.1.0
openai==1.66.2
python-dotenv==1.0.1
gunicorn==21.2.0
# Development dependencies
torch
transformers